# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClassMeta {
  id: string;
  name: string;
  description: string;
  icon: string;
  color: string;
}

export type TalentNodePosition = {
  x: number;
  y: number;
};

export type TalentNodeType =
  (typeof TalentNodeType)[keyof typeof TalentNodeType];

export const TalentNodeType = {
  passive: "passive",
  active: "active",
  choice: "choice",
} as const;

export interface TalentNode {
  id: string;
  name: string;
  description: string;
  maxPoints: number;
  currentPoints: number;
  prerequisites: string[];
  position: TalentNodePosition;
  icon?: string;
  type: TalentNodeType;
}

export interface TalentTree {
  class: string;
  classId: string;
  maxPoints: number;
  color: string;
  nodes: TalentNode[];
}

export interface SaveBuildBody {
  classId: string;
  buildData: string;
  name?: string;
}

export interface SavedBuild {
  id: string;
  classId: string;
  buildData: string;
  name?: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}