# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns metadata for all 21 WoW classes
 * @summary List all classes
 */
export const ListClassesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string(),
  icon: zod.string(),
  color: zod.string(),
});
export const ListClassesResponse = zod.array(ListClassesResponseItem);

/**
 * Returns full talent tree data including all nodes
 * @summary Get talent tree for a class
 */
export const GetClassParams = zod.object({
  classId: zod.coerce.string(),
});

export const GetClassResponse = zod.object({
  class: zod.string(),
  classId: zod.string(),
  maxPoints: zod.number(),
  color: zod.string(),
  nodes: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      description: zod.string(),
      maxPoints: zod.number(),
      currentPoints: zod.number(),
      prerequisites: zod.array(zod.string()),
      position: zod.object({
        x: zod.number(),
        y: zod.number(),
      }),
      icon: zod.string().optional(),
      type: zod.enum(["passive", "active", "choice"]),
    }),
  ),
});

/**
 * Saves a serialized build and returns a shareable ID
 * @summary Save a talent build
 */
export const SaveBuildBody = zod.object({
  classId: zod.string(),
  buildData: zod.string(),
  name: zod.string().optional(),
});

/**
 * Loads a previously saved talent build
 * @summary Load a saved build
 */
export const GetBuildParams = zod.object({
  buildId: zod.coerce.string(),
});

export const GetBuildResponse = zod.object({
  id: zod.string(),
  classId: zod.string(),
  buildData: zod.string(),
  name: zod.string().optional(),
  createdAt: zod.string(),
});