# lib/api-zod/src/generated/types/classMeta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassMeta {
  id: string;
  name: string;
  description: string;
  icon: string;
  color: string;
}