# lib/api-zod/src/generated/types/saveBuildBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveBuildBody {
  classId: string;
  buildData: string;
  name?: string;
}