# lib/api-zod/src/generated/types/savedBuild.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedBuild {
  id: string;
  classId: string;
  buildData: string;
  name?: string;
  createdAt: string;
}