# lib/api-zod/src/generated/types/talentNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { TalentNodePosition } from "./talentNodePosition";
import type { TalentNodeType } from "./talentNodeType";

export interface TalentNode {
  id: string;
  name: string;
  description: string;
  maxPoints: number;
  currentPoints: number;
  prerequisites: string[];
  position: TalentNodePosition;
  icon?: string;
  type: TalentNodeType;
}