# lib/api-zod/src/generated/types/talentTree.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Conquest of Azeroth Talent Calculator API
 * OpenAPI spec version: 0.1.0
 */
import type { TalentNode } from "./talentNode";

export interface TalentTree {
  class: string;
  classId: string;
  maxPoints: number;
  color: string;
  nodes: TalentNode[];
}